pyproject.toml
[tool.poetry]
name = "arcaflow-plugin-kubeconfig"
version = "0.1.0"
description = ""
authors = ["Arcalot Contributors"]
license = "Apache-2.0"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
arcaflow-plugin-sdk = "0.9.0"
kubernetes = "^24.2.0"
PyYAML = ">=5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
